requirements.txt
numpy
opencv-python
ffmpeg-python
loguru
matplotlib
dotenv-python